Don't use a hardcoded election ID for mastership arbitration
